requirements.txt
fastmcp>=0.1.0
jira>=3.5.2
python-dotenv>=1.0.1
